pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ci-agent"
version = "0.1.0"
description = "Competitive Intelligence Agent using the OpenAI Agents SDK and Meta Language"
readme = "README.md"
requires-python = ">=3.11"
authors = [{name = "Your Name", email = "you@example.com"}]
dependencies = [
  "openai-agents>=0.1.0",
  "pydantic>=2.6",
]

[project.optional-dependencies]
ui = [
  "streamlit>=1.36",
]
dev = [
  "ruff>=0.5.0",
  "pytest>=8",
]
google = [
  "google-generativeai>=0.3.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "UP", "B"]
ignore = ["E501"]
fix = true
exclude = ["uv.lock", ".venv", "dist", "build"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.setuptools.packages.find]
where = ["src"]
